% Generated by roxygen2 (4.0.1): do not edit by hand
\name{.cleanFilename}
\alias{.cleanFilename}
\title{This function removes spaces and punctuations from filenames.}
\usage{
.cleanFilename(x)
}
\value{
filename
}
\description{
This function removes spaces and punctuations from filenames.
}
\keyword{internal}

