\docType{methods}
\name{.composeFilename,AbstractMassObject-method}
\alias{.composeFilename,AbstractMassObject-method}
\title{This method creates a filename for a \code{\linkS4class{AbstractMassObject}}
object.}
\usage{
\S4method{.composeFilename}{AbstractMassObject}(x, fileExtension = "csv")
}
\arguments{
  \item{x}{a \code{\linkS4class{AbstractMassObject}}
  object}

  \item{fileExtension}{file type (e.g. "txt", "pdf", ...)}
}
\value{
filename
}
\description{
This method creates a filename for a
\code{\linkS4class{AbstractMassObject}} object.
}
\seealso{
\code{\linkS4class{AbstractMassObject}}
}
\keyword{internal}

