% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export-methods.R
\docType{methods}
\name{exportImzMl,MassSpectrum-method}
\alias{exportImzMl}
\alias{exportImzMl,MassSpectrum-method}
\alias{exportImzMl,list-method}
\title{Export to imzML files}
\usage{
\S4method{exportImzMl}{MassSpectrum}(x, file, force=FALSE, \ldots)

\S4method{exportImzMl}{list}(x, path, force=FALSE, \ldots)
}
\arguments{
\item{x}{a \code{\link[MALDIquant]{MassSpectrum-class}} object or a
\code{list} of \code{\link[MALDIquant]{MassSpectrum-class}} objects.}

\item{file}{\code{character}, file name.}

\item{force}{\code{logical}, If \code{TRUE} the \code{file} would be
overwritten or \code{path} would be created.}

\item{path}{\code{character}, path to directory in which the \code{list} of
\code{\link[MALDIquant]{MassSpectrum-class}} would be exported. If
\code{path} is a single filename all spectra will be exported to a single
imzML file.}

\item{\ldots}{arguments to be passed to internal functions.}
}
\description{
This function exports
\code{\link[MALDIquant]{MassSpectrum-class}} objects into imzML files.
}
\examples{
\dontrun{
library("MALDIquant")
library("MALDIquantForeign")

## get example directory
exampleDirectory <- system.file("exampledata", package="MALDIquantForeign")

## import
s <- importImzMl(file.path(exampleDirectory, "tiny_continuous.imzML"))
## export a list of spectra
exportImzMl(s, file="continuous.imzML")
}
}
\author{
Sebastian Gibb
}
\references{
\url{http://strimmerlab.org/software/maldiquant/}

Schramm T, Hester A, Klinkert I, Both J-P, Heeren RMA, Brunelle A,
Laprévote O, Desbenoit N, Robbe M-F, Stoeckli M, Spengler B, Römpp A
(2012)\cr
imzML - A common data format for the flexible exchange and processing of mass
spectrometry imaging data. \cr
Journal of Proteomics 75 (16):5106-5110. \cr
\url{http://dx.doi.org/10.1016/j.jprot.2012.07.026}
}
\seealso{
\code{\link[MALDIquant]{MassSpectrum-class}}
}

