% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-functions.R
\name{importAnalyze}
\alias{importAnalyze}
\title{Import Analyze 7.5 files}
\usage{
importAnalyze(path, ...)
}
\arguments{
\item{path}{\code{character}, path to directory or file which should be read
in.}

\item{\ldots}{arguments to be passed to
\code{\link[MALDIquantForeign]{import}}.}
}
\value{
a \code{list} of \code{\link[MALDIquant]{MassSpectrum-class}} or
\code{\link[MALDIquant]{MassPeaks-class}} objects (depending on the
\code{centroided} argument).
}
\description{
This function imports files in Analyze 7.5 file format
into \code{\link[MALDIquant]{MassSpectrum-class}} or
\code{\link[MALDIquant]{MassPeaks-class}} objects.
}
\references{
\url{https://strimmerlab.github.io/software/maldiquant/} \cr
 \url{http://www.grahamwideman.com/gw/brain/analyze/formatdoc.htm},
 \url{https://eeg.sourceforge.net/ANALYZE75.pdf}
}
\seealso{
\code{\link[MALDIquant]{MassSpectrum-class}},
\code{\link[MALDIquant]{MassPeaks-class}}
}
\author{
Sebastian Gibb
}
