% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-functions.R
\name{importImzMl}
\alias{importImzMl}
\title{Import imzML files}
\usage{
importImzMl(path, coordinates = NULL, attachOnly = FALSE,
  duplicateFile = TRUE, ...)
}
\arguments{
\item{path}{\code{character}, path to directory or file which should be read
in.}

\item{coordinates}{\code{matrix}, 2 column matrix that contains the x- and
y-coordinates for spectra that should be imported. Other spectra would be
ignored.}

\item{attachOnly}{logical (defaults to \code{FALSE}), whether to attach the dataset via the \code{matter}
package without loading it into memory. See \code{\link[MALDIquant]{MassSpectrumOnDisk-class}}.}

\item{duplicateFile}{logical, when \code{TRUE} (default), creates a temporary copy of the binary \code{ibd}
file in the \code{tempdir} and attaches the \code{\link[MALDIquant]{MassSpectrumOnDisk}} objects to it so 
as not to affect the original \code{ibd} file.}

\item{\ldots}{arguments to be passed to
\code{\link[MALDIquantForeign]{import}}.}
}
\value{
a \code{list} of \code{\link[MALDIquant]{MassSpectrum-class}} or
\code{\link[MALDIquant]{MassPeaks-class}} objects (depending on the
\code{centroided} argument).
}
\description{
This function imports files in imzML file format
into \code{\link[MALDIquant]{MassSpectrum-class}}, 
\code{\link[MALDIquant]{MassSpectrumOnDisk-class}} or
\code{\link[MALDIquant]{MassPeaks-class}} objects.
}
\examples{

library("MALDIquant")
library("MALDIquantForeign")

## get example directory
exampleDirectory <- system.file("exampledata", package="MALDIquantForeign")

## import
s <- importImzMl(file.path(exampleDirectory, "tiny_continuous.imzML"))

## import only spectra for pixel 1,1 and 2,1
s <- importImzMl(file.path(exampleDirectory, "tiny_continuous.imzML"),
                 coordinates = cbind(1:2, c(1, 1)))

}
\references{
\url{http://strimmerlab.org/software/maldiquant/}, \cr\cr
Definition of \code{imzML} format:
\url{http://www.imzml.org/}\cr\cr
\code{"matter"}: Kylie A. Bemis (2018). matter: A framework for rapid prototyping with binary data on disk. R
package version 1.8.0. \url{https://github.com/kuwisdelu/matter}.
}
\seealso{
\code{\link[MALDIquant]{MassSpectrum-class}},
\code{\link[MALDIquant]{MassPeaks-class}}
}
\author{
Sebastian Gibb
}
