% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MALDIquantForeign-package}
\alias{MALDIquantForeign}
\alias{MALDIquantForeign-package}
\title{Import/Export routines for \sQuote{MALDIquant}}
\description{
This package reads and writes different file formats of mass
spectrometry data into/from \sQuote{MALDIquant} objects.
}
\details{
\tabular{ll}{
Package: \tab MALDIquantForeign \cr
License: \tab GPL (>= 3)\cr
URL: \tab https://strimmerlab.github.io/software/maldiquant/\cr
}
}
\references{
\url{https://strimmerlab.github.io/software/maldiquant/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://strimmerlab.github.io/software/maldiquant/}
  \item \url{https://github.com/sgibb/MALDIquantForeign/}
  \item Report bugs at \url{https://github.com/sgibb/MALDIquantForeign/issues/}
}

}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\keyword{package}
