% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parseCiphergenXml.R
\name{.parseCiphergenXml}
\alias{.parseCiphergenXml}
\title{Parse Ciphergen XML files.}
\usage{
.parseCiphergenXml(file, ...)
}
\arguments{
\item{file}{\code{character}, path to Ciphergen XML file}

\item{verbose}{\code{logical}, verbose output?}
}
\value{
Returns a list with metadata and spectra.
}
\description{
This function parses Ciphergen XML files.
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\keyword{internal}

