% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filename-functions.R
\name{.cleanFilename}
\alias{.cleanFilename}
\title{This function removes spaces and punctuations from filenames.}
\usage{
.cleanFilename(x)
}
\value{
filename
}
\description{
This function removes spaces and punctuations from filenames.
}
\keyword{internal}

