\docType{methods}
\name{.composeFilename}
\alias{.composeFilename}
\alias{.composeFilename,AbstractMassObject-method}
\alias{.composeFilename,list-method}
\title{This method creates a filename for a \code{\linkS4class{AbstractMassObject}}
object.}
\arguments{
  \item{x}{a \code{\linkS4class{AbstractMassObject}}
  object}

  \item{fileExtension}{file type (e.g. "txt", "pdf", ...)}
}
\value{
  filename
}
\description{
  This method creates a filename for a
  \code{\linkS4class{AbstractMassObject}} object.
}
\seealso{
  \code{\linkS4class{AbstractMassObject}}
}
\keyword{internal}

