% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MALDIquantForeign-package}
\alias{MALDIquantForeign-package}
\title{Import/Export routines for \sQuote{MALDIquant}}
\description{
This package reads and writes different file formats of mass
spectrometry data into/from \sQuote{MALDIquant} objects.
}
\details{
\tabular{ll}{
Package: \tab MALDIquantForeign \cr
License: \tab GPL (>= 3)\cr
URL: \tab http://strimmerlab.org/software/maldiquant/\cr
}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\references{
\url{http://strimmerlab.org/software/maldiquant/}
}
\keyword{package}

