\docType{package}
\name{MALDIquantForeign-package}
\alias{MALDIquantForeign-package}
\title{Import/Export routines for MALDIquant}
\description{
  This package allows to read and write different file
  formats of mass spectrometry data into/from MALDIquant
  objects.
}
\details{
  \tabular{ll}{ Package: \tab MALDIquantForeign \cr
  Version: \tab 0.1.2\cr Date: \tab 2012-12-11\cr License:
  \tab GPL (>= 3)\cr URL: \tab
  http://strimmerlab.org/software/maldiquant/\cr }
}
\author{
  Sebastian Gibb <\email{mail@sebastiangibb.de}>
}
\references{
  \url{http://strimmerlab.org/software/maldiquant/}
}
\keyword{package}

