% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileFormats.R
\name{supportedFileFormats}
\alias{supportedFileFormats}
\title{Supported file formats}
\usage{
supportedFileFormats()
}
\value{
a \code{list} with two named elements (\code{import} and
\code{export}) containing a \code{character} vector of supported file types.
}
\description{
This function prints all file formats supported by
\code{\link{MALDIquantForeign-package}}.
}
\details{
\subsection{Import}{

\tabular{ll}{
 txt \tab \code{\link[MALDIquantForeign]{importTxt}} \cr
 tab \tab \code{\link[MALDIquantForeign]{importTab}} \cr
 csv \tab \code{\link[MALDIquantForeign]{importCsv}} \cr
 fid \tab \code{\link[MALDIquantForeign]{importBrukerFlex}} \cr
 ciphergen \tab \code{\link[MALDIquantForeign]{importCiphergenXml}} \cr
 mzXML \tab \code{\link[MALDIquantForeign]{importMzXml}} \cr
 mzML \tab \code{\link[MALDIquantForeign]{importMzMl}} \cr
 imzML \tab \code{\link[MALDIquantForeign]{importImzMl}} \cr
 analyze \tab \code{\link[MALDIquantForeign]{importAnalyze}} \cr
 cdf \tab \code{\link[MALDIquantForeign]{importCdf}} \cr
 msd \tab \code{\link[MALDIquantForeign]{importMsd}} \cr
}
}

\subsection{Export}{

\tabular{ll}{
 tab \tab \code{\link[MALDIquantForeign]{exportTab}} \cr
 csv \tab \code{\link[MALDIquantForeign]{exportCsv}} \cr
 imzML \tab \code{\link[MALDIquantForeign]{exportImzMl}} \cr
 msd \tab \code{\link[MALDIquantForeign]{exportMsd}} \cr
 mzML \tab \code{\link[MALDIquantForeign]{exportMzMl}} \cr
}
}
}
\examples{
library("MALDIquantForeign")

supportedFileFormats()

}
\author{
Sebastian Gibb
}
\references{
\url{http://strimmerlab.org/software/maldiquant/}
}
\seealso{
\code{\link[MALDIquantForeign]{export}},
 \code{\link[MALDIquantForeign]{import}}
}

